working progress change message parsing

+++ app/src/main/java/certh/hit/cmobile/model/UserMessage.kt
@@ -5,8 +5,8 @@ package certh.hit.cmobile.model
  */
 open class UserMessage {
     var topic: Topic? = null
-    var latitude: Double? = null
-    var longitude: Double? = null
+    var latitude: Int? = null
+    var longitude: Int? = null
 
 
 }
